Moved more external libs into libraries project, renamed from openexr project. Fixed a few compile warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2305 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Hilbert.cpp
@@ -94,7 +94,7 @@ void Hilbert::test()
 		indices
 	);
 
-	for(int i=0; i<indices.size(); ++i)
+	for(size_t i=0; i<indices.size(); ++i)
 	{
 		conPrint("index " + ::toString(i) + ": " + toString(indices[i].first) + ", " + toString(indices[i].second));
 	}
